+++ core/src/main/java/org/apache/mahout/classifier/bayes/common/BayesTfIdfDriver.java
@@ -20,7 +20,7 @@ package org.apache.mahout.classifier.bayes.common;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.DefaultStringifier;
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.JobClient;
 import org.apache.hadoop.mapred.JobConf;
@@ -45,7 +45,8 @@ public class BayesTfIdfDriver {
    * Takes in two arguments:
    * <ol>
    * <li>The input {@link org.apache.hadoop.fs.Path} where the input documents live</li>
-   * <li>The output {@link org.apache.hadoop.fs.Path} where to write the interim files as a {@link org.apache.hadoop.io.SequenceFile}</li>
+   * <li>The output {@link org.apache.hadoop.fs.Path} where to write the interim files as a
+   *  {@link org.apache.hadoop.io.SequenceFile}</li>
    * </ol>
    * @param args The args
    */
@@ -68,7 +69,7 @@ public class BayesTfIdfDriver {
     
 
     conf.setOutputKeyClass(Text.class);
-    conf.setOutputValueClass(FloatWritable.class);
+    conf.setOutputValueClass(DoubleWritable.class);
     
     SequenceFileInputFormat.addInputPath(conf, new Path(output + "/trainer-termDocCount"));
     SequenceFileInputFormat.addInputPath(conf, new Path(output + "/trainer-wordFreq"));
@@ -94,14 +95,14 @@ public class BayesTfIdfDriver {
 
     Path interimFile = new Path(output+"/trainer-docCount/part-*");
 
-    Map<String,Float> labelDocumentCounts= reader.readLabelDocumentCounts(dfs, interimFile, conf);
+    Map<String,Double> labelDocumentCounts= reader.readLabelDocumentCounts(dfs, interimFile, conf);
 
-    DefaultStringifier<Map<String,Float>> mapStringifier =
-        new DefaultStringifier<Map<String,Float>>(conf,GenericsUtil.getClass(labelDocumentCounts));
+    DefaultStringifier<Map<String,Double>> mapStringifier =
+        new DefaultStringifier<Map<String,Double>>(conf,GenericsUtil.getClass(labelDocumentCounts));
 
     String labelDocumentCountString = mapStringifier.toString(labelDocumentCounts);
     log.info("Counts of documents in Each Label");
-    Map<String,Float> c = mapStringifier.fromString(labelDocumentCountString);
+    Map<String,Double> c = mapStringifier.fromString(labelDocumentCountString);
     log.info("{}", c);
 
     conf.set("cnaivebayes.labelDocumentCounts", labelDocumentCountString);
