+++ core/src/main/java/org/apache/mahout/classifier/cbayes/CBayesClassifier.java
@@ -48,8 +48,8 @@ public class CBayesClassifier implements Classifier{
     PriorityQueue pq = new ClassifierResultPriorityQueue(numResults);
     ClassifierResult tmp;
     for (String category : categories){
-      float prob = documentProbability(model, category, document);
-      if (prob < 0) {
+      double prob = documentProbability(model, category, document);
+      if (prob < 0.0) {
         tmp = new ClassifierResult(category, prob);
         pq.insert(tmp);
       }
@@ -75,11 +75,11 @@ public class CBayesClassifier implements Classifier{
    */
   public ClassifierResult classify(Model model, String[] document, String defaultCategory) {
     ClassifierResult result = new ClassifierResult(defaultCategory);
-    float min = 0.0f;
+    double min = 0.0;
     Collection<String> categories = model.getLabels();
 
     for (String category : categories) {
-      float prob = documentProbability(model, category, document);
+      double prob = documentProbability(model, category, document);
       if (prob < min) {
         min = prob;
         result.setLabel(category);
@@ -99,8 +99,8 @@ public class CBayesClassifier implements Classifier{
    * @return The probability
    * @see Model#FeatureWeight(String, String)
    */
-  public float documentProbability(Model model, String label, String[] document) {
-    float result = 0.0f;
+  public double documentProbability(Model model, String label, String[] document) {
+    double result = 0.0;
     Map<String, Integer> wordList = new HashMap<String, Integer>(1000);
     for (String word : document) {
       if (wordList.containsKey(word)) {
@@ -129,8 +129,8 @@ public class CBayesClassifier implements Classifier{
       ClassifierResult cr1 = (ClassifierResult) a;
       ClassifierResult cr2 = (ClassifierResult) b;
 
-      float score1 = cr1.getScore();
-      float score2 = cr2.getScore();
+      double score1 = cr1.getScore();
+      double score2 = cr2.getScore();
       return score1 == score2 ? cr1.getLabel().compareTo(cr2.getLabel()) < 0 : score1 < score2;
     }
   }
